Wire up the full 4-task FreeRTOS architecture in the main sketch
Task 1 (alta): captura I2S continua com buffer duplo (int16, PCM),
sincronizado por 2 semaforos binarios (nunca sobrescreve um buffer que a
Task 2 ainda esta lendo).
Task 2 (media): extracao de features, libera o semaforo do buffer assim
que termina de le-lo, envia features por fila.
Task 3 (baixa): forward-pass do modelo + threshold de confianca +
cooldown entre comandos, envia comando por fila.
Task 4 (baixa): aciona o servo correspondente + LED/buzzer, atualiza o
log de latencia (unico recurso protegido por mutex, ja que nao passa
pelas filas).

Reduz uso de RAM: buffers de captura de float (64KB cada) para int16 PCM
(32KB cada), e elimina uma copia intermediaria de 64KB no
feature_extraction.cpp (normalizacao de ganho aplicada por frame, nao
mais materializada como um array separado do tamanho do clipe inteiro).
Sem essas mudancas o link falhava por faltar ~98KB de RAM; com elas,
compila usando 29% da RAM e 25% da flash (verificado com arduino-cli
compile --fqbn esp32:esp32:esp32, real, nao so revisao de codigo).

+++ firmware/dino_voice_controller/feature_extraction.h
@@ -4,11 +4,14 @@
 // firmware/dsp_tables.h, gerado por training/export_c_dsp.py).
 #pragma once
 
+#include <stdint.h>
+
 #include "dsp_tables.h"
 
 #define FEATURE_DIM (1 + 1 + 1 + DSP_N_MFCC)  // RMS + ZCR + centroide + MFCC
 
-// signal deve ter exatamente DSP_CLIP_LENGTH (16000) amostras normalizadas
-// em [-1, 1]. outFeatures recebe FEATURE_DIM valores, na ordem
+// signal deve ter exatamente DSP_CLIP_LENGTH (16000) amostras PCM de 16 bits
+// (mesma convenção de um WAV int16 — valor float = amostra / 32768.0).
+// outFeatures recebe FEATURE_DIM valores, na ordem
 // [RMS, ZCR, centroide, mfcc_0..mfcc_10] — mesma ordem de training/features.py.
-void extractFeatures(const float *signal, float outFeatures[FEATURE_DIM]);
+void extractFeatures(const int16_t *signal, float outFeatures[FEATURE_DIM]);
